Reverted implicit orphan-ignoring lfsr_mtree_pathlookup
This did not have any benefit code-size wise, and while it may be nice
for lfsr_mtree_pathlookup to take care of orphans, leaving it up to the
upper-layers is both simpler and makes orphan behavior explicit in all
functions.

We don't really have that many functions without special orphan
behavior anyways.

           code          stack          ctx
  before: 38124 (+0.0%)   2624 (+0.0%)  752 (+0.0%)
  after:  38120 (+0.0%)   2624 (+0.0%)  752 (+0.0%)