Made did generation in lfsr_mkdir deterministic
Previously, we were using the checksum of the full path to generate
dids. This mostly works, but means that different paths can end up with
different dids for the same directory:

- crc32c("a/b")       => 0xfb0b40a3
- crc32c("/a/c/../b") => 0x223404f6

Not the end of the world, but this is the stuff heisenbugs are made of.

Now we instead checksum only the file name, which doesn't have this
problem, and xor with our parent's did to prevent collisions between
same-named files in different directories:

  did = parent_did xor crc32c(name)

As an extra plus, this should also trivially work for the theoretical
lfsr_mkdirat function.

Code size unchanged, which is humorous but not surprising:

           code          stack          ctx
  before: 38120           2624          752
  after:  38120 (+0.0%)   2624 (+0.0%)  752 (+0.0%)