Tweaked LFSR_DATA_SLICE/TRUNCATE/FRUNCATE to return lvalues
This is to be consistent with other LFSR_DATA_* constructors. The code
is also a bit more readable when all LFSR_DATA_* constructors are
capitalized.

Not really sure why, but this saved a bit of stack? Probably just
compiler noise:

             code          stack          ctx
  before:   38144           2616          752
  after:    38148 (+0.0%)   2608 (-0.3%)  752 (+0.0%)

I also explored adding in-place slice/truncate/fruncate functions as
well, but the impact on stack usage was REALLY BAD:

             code          stack          ctx
  by-value: 38148           2608          752
  in-place: 38144 (-0.0%)   2688 (+3.1%)  752 (+0.0%)

I think maybe because the in-place functions end up with too many
pointers for the compiler to make safe assumptions about compound-
literal lifetimes?