Dropped LFSR_DATA_NAME cat-chopping hack
Unfortunately this is undefined behavior.

As far as I can tell, there's no well-defined way in C to express that
we don't need a full struct allocation.

Curiously this ended up saving code? I guess because of better compiler
assumptions when using the correct types. This hack was supposed to save
stack, but it's possible the single saved word was lost due to alignment/
measurement noise:

           code          stack          ctx
  before: 38060           2608          752
  after:  38036 (-0.1%)   2608 (+0.0%)  752 (+0.0%)